+++ openpeer/services/IHelper.h
@@ -76,7 +76,7 @@ namespace openpeer
 
       static String randomString(UINT lengthInChars);
 
-      static SecureByteBlockPtr random(UINT lengthInBytes);
+      static SecureByteBlockPtr random(size_t lengthInBytes);
 
       static ULONG random(ULONG minValue, ULONG maxValue);
 
@@ -93,12 +93,12 @@ namespace openpeer
       static SecureByteBlockPtr convertToBuffer(const std::string &input);
       static SecureByteBlockPtr convertToBuffer(
                                                 const BYTE *buffer,
-                                                ULONG bufferLengthInBytes
+                                                size_t bufferLengthInBytes
                                                 );
 
       static String convertToBase64(
                                     const BYTE *buffer,
-                                    ULONG bufferLengthInBytes
+                                    size_t bufferLengthInBytes
                                     );
 
       static String convertToBase64(const SecureByteBlock &input);
@@ -111,7 +111,7 @@ namespace openpeer
 
       static String convertToHex(
                                  const BYTE *buffer,
-                                 ULONG bufferLengthInBytes,
+                                 size_t bufferLengthInBytes,
                                  bool outputUpperCase = false
                                  );
 
@@ -235,7 +235,7 @@ namespace openpeer
 
       static String getDebugString(
                                    const BYTE *buffer,
-                                   ULONG bufferSizeInBytes,
+                                   size_t bufferSizeInBytes,
                                    ULONG bytesPerGroup = 4,
                                    ULONG maxLineLength = 160
                                    );
